create app/modules.go
